+++ app/src/main/java/com/poupa/vinylmusicplayer/ui/activities/AlbumDetailActivity.java
@@ -136,27 +136,6 @@ public class AlbumDetailActivity extends AbsSlidingMusicPanelActivity implements
     }
 
     @Override
-    protected void onStart() {
-        super.onStart();
-
-        Discography.getInstance().addChangedListener(onDiscographyChanged);
-    }
-
-    @Override
-    protected void onStop() {
-        Discography.getInstance().removeChangedListener(onDiscographyChanged);
-
-        super.onStop();
-    }
-
-    @Override
-    protected void onRestart() {
-        super.onRestart();
-
-        reload();
-    }
-
-    @Override
     protected View createContentView() {
         return wrapSlidingMusicPanel(R.layout.activity_album_detail);
     }
@@ -267,7 +246,8 @@ public class AlbumDetailActivity extends AbsSlidingMusicPanelActivity implements
         });
     }
 
-    private void reload() {
+    @Override
+    protected void reload() {
         LoaderManager.getInstance(this).restartLoader(LOADER_ID, getIntent().getExtras(), this);
     }
