+++ app/src/main/java/com/poupa/vinylmusicplayer/ui/activities/ArtistDetailActivity.java
@@ -154,24 +154,6 @@ public class ArtistDetailActivity extends AbsSlidingMusicPanelActivity implement
     }
 
     @Override
-    protected void onStart() {
-        super.onStart();
-        Discography.getInstance().addChangedListener(onDiscographyChanged);
-    }
-
-    @Override
-    protected void onStop() {
-        Discography.getInstance().removeChangedListener(onDiscographyChanged);
-        super.onStop();
-    }
-
-    @Override
-    protected void onRestart() {
-        super.onRestart();
-        reload();
-    }
-
-    @Override
     protected View createContentView() {
         return wrapSlidingMusicPanel(R.layout.activity_artist_detail);
     }
@@ -228,7 +210,8 @@ public class ArtistDetailActivity extends AbsSlidingMusicPanelActivity implement
         this.usePalette = usePalette;
     }
 
-    private void reload() {
+    @Override
+    protected void reload() {
         LoaderManager.getInstance(this).restartLoader(LOADER_ID, getIntent().getExtras(), this);
     }
