+++ app/src/main/java/com/poupa/vinylmusicplayer/ui/activities/MainActivity.java
@@ -370,6 +370,9 @@ public class MainActivity extends AbsSlidingMusicPanelActivity {
         }
     }
 
+    @Override
+    protected void reload() {}
+
     public interface MainActivityFragmentCallbacks {
         boolean handleBackPress();
     }
