+++ app/src/main/java/com/poupa/vinylmusicplayer/ui/activities/PlaylistDetailActivity.java
@@ -95,24 +95,6 @@ public class PlaylistDetailActivity extends AbsSlidingMusicPanelActivity impleme
     }
 
     @Override
-    protected void onStart() {
-        super.onStart();
-        Discography.getInstance().addChangedListener(onDiscographyChanged);
-    }
-
-    @Override
-    protected void onStop() {
-        Discography.getInstance().removeChangedListener(onDiscographyChanged);
-        super.onStop();
-    }
-
-    @Override
-    protected void onRestart() {
-        super.onRestart();
-        reload();
-    }
-
-    @Override
     protected View createContentView() {
         return wrapSlidingMusicPanel(R.layout.activity_playlist_detail);
     }
@@ -298,7 +280,8 @@ public class PlaylistDetailActivity extends AbsSlidingMusicPanelActivity impleme
             adapter.swapDataSet(new ArrayList<>());
     }
 
-    private void reload() {
+    @Override
+    protected void reload() {
         LoaderManager.getInstance(this).restartLoader(LOADER_ID, null, this);
     }
